+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "ls-interactive"
 description = "Interactive ls command"
-version = "1.1.2"
+version = "1.2.0"
 authors = ["Araxeus <oaeben@gmail.com>"]
 homepage = "https://github.com/Araxeus/ls-interactive"
 license = "MIT"
