+++ src/utils.rs
@@ -1,7 +1,14 @@
-use std::{fmt::Display, fs, mem, panic};
+use crate::structs::Entry;
 
-use console::style;
+use console::{style, Term};
+use dialoguer::{theme::ColorfulTheme, Select};
 use lnk::ShellLink;
+
+use std::{fmt::Display, fs, mem, panic, process};
+
+pub fn err<S: Display>(msg: S) {
+    println!("{} {}", style("Error").red(), msg);
+}
 
 pub fn resolve_lnk(path: &String) -> String {
     panic::set_hook(Box::new(|_info| {}));
@@ -26,6 +33,20 @@ pub fn resolve_lnk(path: &String) -> String {
     }
 }
 
-pub fn err<S: Display>(msg: S) {
-    println!("{} {}", style("Error").red(), msg);
+// dialoguer select from a list of choices
+// returns the index of the selected choice
+pub fn display_choices(items: &[Entry], path: &str) -> usize {
+    match Select::with_theme(&ColorfulTheme::default())
+        .with_prompt(&path[4..])
+        .report(false)
+        .items(items)
+        .default(0)
+        .interact_on_opt(&Term::stderr())
+        .ok()
+        .unwrap()
+    {
+        Some(index) => index,
+        // exit process if none
+        None => process::exit(0),
+    }
 }
